moved routes specific code to a new module
